# app/routes/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Content-X (Groot) API",
    "description": "Schema-driven CMS API for the Fynd Platform extension **groot**. Routes under `/api` are served through the FDK extension router (`fdk-extension-javascript`); clients must have a valid extension / platform session (cookies and OAuth flow). `company_id` and `application_id` are taken from session, query, or body per `request-context.helper.js`. Published CMS instance routes use `router.param('slug', loadDefinition)` and require a **published** definition for that slug.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as the extension (paths include `/api/v1/...`)"
    }
  ],
...
  "paths": {
...
    "/api/v1/cms/definitions": {
      "get": {
        "tags": ["Definitions"],
        "summary": "List object definitions for the current company + application",
        "operationId": "listObjectDefinitions",
        "parameters": [
          {
            "name": "application_id",
            "in": "query",
            "description": "Target application (also from FDK session / body in some clients).",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of definitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ObjectDefinition" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Definitions"],
        "summary": "Create object definition",
        "operationId": "createObjectDefinition",
        "parameters": [
          {
            "name": "application_id",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ObjectDefinitionCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectDefinition" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/cms/definitions/{definitionId}": {
      "get": {
        "tags": ["Definitions"],
        "summary": "Get object definition by id",
        "operationId": "getObjectDefinition",
        "parameters": [
          { "$ref": "#/components/parameters/DefinitionId" },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectDefinition" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Definitions"],
        "summary": "Update object definition (resets to DRAFT, bumps version)",
        "operationId": "updateObjectDefinition",
        "parameters": [
          { "$ref": "#/components/parameters/DefinitionId" },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ObjectDefinitionUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectDefinition" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "delete": {
        "tags": ["Definitions"],
        "summary": "Delete object definition and tenant-scoped instances",
        "operationId": "deleteObjectDefinition",
        "parameters": [
          { "$ref": "#/components/parameters/DefinitionId" },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "Definition is referenced by custom_object fields on other definitions",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReferencedError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/cms/definitions/{definitionId}/deletion-impact": {
      "get": {
        "tags": ["Definitions"],
        "summary": "Preview deletion impact (entries count, referencing definitions)",
        "operationId": "getDefinitionDeletionImpact",
        "parameters": [
          { "$ref": "#/components/parameters/DefinitionId" },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeletionImpact" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/cms/definitions/{definitionId}/publish": {
      "post": {
        "tags": ["Definitions"],
        "summary": "Publish definition (creates version snapshot, invalidates cache)",
        "operationId": "publishObjectDefinition",
        "parameters": [
          { "$ref": "#/components/parameters/DefinitionId" },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectDefinition" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/cms/definitions/{definitionId}/versions": {
      "get": {
        "tags": ["Definitions"],
        "summary": "List version history",
        "operationId": "listObjectDefinitionVersions",
        "parameters": [
          { "$ref": "#/components/parameters/DefinitionId" },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ObjectDefinitionVersion" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "DefinitionId": {
        "name": "definitionId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
    "schemas": {
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
      "ReferencedError": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "referenced_by": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "slug": { "type": "string" },
                "name": { "type": "string" }
              }
            }
          }
        }
      },
      "FieldSchema": {
        "type": "object",
        "description": "Single field in an object definition",
        "properties": {
          "name": { "type": "string" },
          "label": { "type": "string" },
          "field_type": {
            "type": "string",
            "enum": [
              "text",
              "long_text",
              "number",
              "file",
              "date",
              "date_time",
              "url",
              "json",
              "html",
              "multiple_values",
              "dropdown",
              "checkbox",
              "radio",
              "product",
              "collection",
              "custom_object"
            ]
          },
          "required": { "type": "boolean" },
          "regex_pattern": { "type": "string" },
          "is_filterable": { "type": "boolean" },
          "options": { "type": "array", "items": { "type": "string" } },
          "default_value": {},
          "multiple": { "type": "boolean" },
          "relation": { "type": "string", "enum": ["none", "product", "collection"] },
          "placeholder": { "type": "string" },
          "ref_definition_slug": { "type": "string" },
          "file_config": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": true
      },
      "ObjectDefinition": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "company_id": { "type": "string" },
          "application_id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "usage_mode": { "type": "string", "enum": ["standalone", "embedded_only"] },
          "status": { "type": "string", "enum": ["DRAFT", "PUBLISHED"] },
          "version": { "type": "integer" },
          "fields": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FieldSchema" }
          },
          "published_at": { "type": "string", "format": "date-time" },
          "created_by": { "type": "string" },
          "updated_by": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": true
      },
      "ObjectDefinitionCreate": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string", "description": "Optional; derived from name if omitted" },
          "description": { "type": "string" },
          "usage_mode": { "type": "string", "enum": ["standalone", "embedded_only"] },
          "fields": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FieldSchema" }
          }
        }
      },
      "ObjectDefinitionUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "usage_mode": { "type": "string", "enum": ["standalone", "embedded_only"] },
          "fields": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FieldSchema" }
          }
        }
      },
      "ObjectDefinitionVersion": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "company_id": { "type": "string" },
          "application_id": { "type": "string" },
          "definition_id": { "type": "string" },
          "version": { "type": "integer" },
          "status": { "type": "string" },
          "snapshot": { "type": "object", "additionalProperties": true },
          "comment": { "type": "string" },
          "created_by": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": true
      },
      "DeletionImpact": {
        "type": "object",
        "properties": {
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "entry_count": { "type": "integer" },
          "referenced_by": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "slug": { "type": "string" },
                "name": { "type": "string" }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      },
...
    }
  }
}